begin battle refactor
